dbsp: Fix shutdown of exchange operator.
The exchange operator created a Tokio task that listened for connections
from other nodes.  This didn't get killed when the Exchange was dropped.
This commit fixes that problem by encapsulating the task in a new
`ExchangeListener` that holds a Tokio `DropGuard` for a `CancellationToken`
to cancel the task.

There was the additional problem of a circular reference: `Runtime` holds
a reference to `Exchange`, which in turn held a reference back to
`Runtime`.  This commit also fixes that problem by avoiding the reference
from `Exchange` back to `Runtime`.

Signed-off-by: Ben Pfaff <blp@feldera.com>